feat: sdk-core exmaple and improve network-exmaple
