+++ apps/sdk-nextjs-integration/src/components/header/index.tsx
@@ -1,0 +1,1 @@
+export * from './header';
